Fix compatibility with Sphinx 7.3 and Sphinx Gallery 0.16
